fix: enable event-based profession and zone type registration
Custom registration events now properly implement IModBusEvent and fire
on the mod event bus during FMLCommonSetupEvent. Event handlers moved
from separate classes into registry classes to ensure classloading.

Resolves #132
