+++ src/main/java/org/sosly/villagetale/event/RegisterProfessionsEvent.java
@@ -1,10 +1,11 @@
 package org.sosly.villagetale.event;
 
 import net.minecraftforge.eventbus.api.Event;
+import net.minecraftforge.fml.event.IModBusEvent;
 import org.sosly.villagetale.api.IProfession;
 import org.sosly.villagetale.profession.ProfessionRegistry;
 
-public class RegisterProfessionsEvent extends Event {
+public class RegisterProfessionsEvent extends Event implements IModBusEvent {
     private final ProfessionRegistry registry;
 
     public RegisterProfessionsEvent(ProfessionRegistry professions) {
