+++ src/main/java/org/sosly/villagetale/event/RegisterZoneTypesEvent.java
@@ -2,10 +2,11 @@ package org.sosly.villagetale.event;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.eventbus.api.Event;
+import net.minecraftforge.fml.event.IModBusEvent;
 import org.sosly.villagetale.api.IZoneType;
 import org.sosly.villagetale.zone.ZoneRegistry;
 
-public class RegisterZoneTypesEvent extends Event {
+public class RegisterZoneTypesEvent extends Event implements IModBusEvent {
     private final ZoneRegistry registry;
 
     public RegisterZoneTypesEvent(ZoneRegistry zoneTypes) {
