--- src/main/java/org/sosly/villagetale/profession/Professions.java
@@ -1,29 +1,0 @@
-package org.sosly.villagetale.profession;
-
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import org.sosly.villagetale.event.RegisterProfessionsEvent;
-import org.sosly.villagetale.profession.professions.Butcher;
-import org.sosly.villagetale.profession.professions.Carpenter;
-import org.sosly.villagetale.profession.professions.Commoner;
-import org.sosly.villagetale.profession.professions.Cook;
-import org.sosly.villagetale.profession.professions.Farmer;
-import org.sosly.villagetale.profession.professions.Fisher;
-import org.sosly.villagetale.profession.professions.Forester;
-import org.sosly.villagetale.profession.professions.Herder;
-import org.sosly.villagetale.profession.professions.Tanner;
-
-//@Mod.EventBusSubscriber(modid = VillageTale.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
-public class Professions {
-    @SubscribeEvent
-    public static void onRegisterProfessions(RegisterProfessionsEvent event) {
-        event.register(new Butcher());
-        event.register(new Carpenter());
-        event.register(new Commoner());
-        event.register(new Cook());
-        event.register(new Farmer());
-        event.register(new Fisher());
-        event.register(new Forester());
-        event.register(new Herder());
-        event.register(new Tanner());
-    }
-}
